+++ src/IPUtils.php
@@ -44,7 +44,7 @@ class IPUtils {
 	 */
 	const RE_IP_PREFIX = '(3[0-2]|[12][0-9]|[0-9])';
 	/** @private */
-	const RE_IP_RANGE = '(' . self::RE_IP_ADD . '\/' . self::RE_IP_PREFIX . '|' . self::RE_IP_ADD . '\-' . self::RE_IP_ADD . ')';
+	const RE_IP_RANGE = '(' . self::RE_IP_ADD . '\/' . self::RE_IP_PREFIX . '|' . self::RE_IP_ADD . ' ?\- ?' . self::RE_IP_ADD . ')';
 
 	/**
 	 * An IPv6 address is made up of 8 words (each x0000 to xFFFF).
@@ -76,7 +76,7 @@ class IPUtils {
 	 * An IPv6 range is an IP address and a prefix (d1 to d128)
 	 * @private
 	 */
-	const RE_IPV6_RANGE = '(' . self::RE_IPV6_ADD . '\/' . self::RE_IPV6_PREFIX . '|' . self::RE_IPV6_ADD . '\-' . self::RE_IPV6_ADD . ')';
+	const RE_IPV6_RANGE = '(' . self::RE_IPV6_ADD . '\/' . self::RE_IPV6_PREFIX . '|' . self::RE_IPV6_ADD . ' ?\- ?' . self::RE_IPV6_ADD . ')';
 	/**
 	 * For IPv6 canonicalization (NOT for strict validation; these are quite lax!)
 	 * @private